lesson 43 added

+++ src/homeworks/homework_42/Task1/Sportsman.java
@@ -5,9 +5,9 @@ import java.util.Comparator;
 public class Sportsman implements Comparable<Sportsman> {
     private String name;
     private int age;
-    private int score;
+    private double score;
 
-    public Sportsman(String name, int age, int score) {
+    public Sportsman(String name, int age, double score) {
         this.name = name;
         this.age = age;
         this.score = score;
@@ -43,7 +43,7 @@ public class Sportsman implements Comparable<Sportsman> {
         this.age = age;
     }
 
-    public int getScore() {
+    public double getScore() {
         return score;
     }
